backend/requirements.txt
absl-py==1.3.0
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
asyncio==3.4.3
attrs==22.1.0
automationhat==0.2.0
beautifulsoup4==4.7.1
biopython==1.80
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
CacheControl==0.12.12
cachetools==5.2.0
Cap1xxx==0.1.3
certifi==2018.8.24
cffi==1.15.1
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
cycler==0.11.0
dnspython==2.2.1
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
Flask-Cors==3.0.10
fonttools==4.38.0
fourletterphat==0.1.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
gpiozero==1.6.2
grpcio==1.51.0
grpcio-status==1.51.0
html5lib==1.0.1
httplib2==0.21.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.4.4
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
matplotlib==3.5.3
mccabe==0.6.1
mediapipe-rpi4==0.8.8
microdotphat==0.2.1
mock==1.0.1
mote==0.0.4
motephat==0.0.3
msgpack==1.0.4
mypy==0.670
mypy-extensions==0.4.1
numpy==1.21.6
oauthlib==2.1.0
olefile==0.46
opencv-contrib-python==4.1.0.25
packaging==21.3
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pymongo==4.3.3
pyOpenSSL==19.0.0
pyparsing==3.0.9
pysam==0.20.0
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
python-dateutil==2.8.2
pyxdg==0.25
rainbowhat==0.1.0
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.9
RTIMULib==7.2.1
RUST==0.1.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.4.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
ssh-import-id==5.7
thonny==3.3.10
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
typing-extensions==4.4.0
unicornhathd==0.0.4
uritemplate==4.1.1
urllib3==1.24.1
webencodings==0.5.1
websockets==10.4
Werkzeug==0.14.1
wrapt==1.10.11
